Refactor integration benchmarks
